move the vmm-building parts from controller to builder
