+++ src/vmm/src/lib.rs
@@ -91,8 +91,6 @@ use vstate::{KvmContext, Vcpu, Vm};
 
 pub use error::{ErrorKind, StartMicrovmError, VmmActionError};
 
-const WRITE_METRICS_PERIOD_SECONDS: u64 = 60;
-
 /// Success exit code.
 pub const FC_EXIT_CODE_OK: u8 = 0;
 /// Generic error exit code.
@@ -395,186 +393,6 @@ pub struct VcpuConfig {
     pub ht_enabled: bool,
     /// CPUID template to use.
     pub cpu_template: Option<CpuFeaturesTemplate>,
-}
-
-/// Encapsulates configuration parameters for a `VmmBuilderz`.
-pub struct VmmBuilderzConfig {
-    /// The guest memory object for this VM.
-    pub guest_memory: GuestMemory,
-    /// The guest physical address of the execution entry point.
-    pub entry_addr: GuestAddress,
-    /// Base kernel command line contents.
-    pub kernel_cmdline: KernelCmdline,
-    /// vCPU configuration paramters.
-    pub vcpu_config: VcpuConfig,
-    /// Seccomp filtering level.
-    pub seccomp_level: u32,
-}
-
-/// Helps build a Vmm.
-pub struct VmmBuilderz {
-    vmm: Vmm,
-    vcpus: Vec<Vcpu>,
-}
-
-impl VmmBuilderz {
-    /// Create a new VmmBuilderz.
-    pub fn new(
-        epoll_context: &mut EpollContext,
-        config: VmmBuilderzConfig,
-    ) -> std::result::Result<Self, VmmActionError> {
-        let write_metrics_event_fd = TimerFd::new_custom(ClockId::Monotonic, true, true)
-            .map_err(Error::TimerFd)
-            .map_err(StartMicrovmError::Internal)?;
-
-        let event_manager = EventManager::new()
-            .map_err(Error::EventManager)
-            .map_err(StartMicrovmError::Internal)?;
-
-        epoll_context
-            .add_epollin_event(&event_manager, EpollDispatch::PollyEvent)
-            .expect("Cannot cascade EventManager from epoll_context");
-
-        epoll_context
-            .add_epollin_event(
-                // non-blocking & close on exec
-                &write_metrics_event_fd,
-                EpollDispatch::WriteMetrics,
-            )
-            .expect("Cannot add write metrics TimerFd to epoll.");
-
-        let kvm = KvmContext::new()
-            .map_err(Error::KvmContext)
-            .map_err(StartMicrovmError::Internal)?;
-
-        let mut vm = Vm::new(kvm.fd())
-            .map_err(Error::Vm)
-            .map_err(StartMicrovmError::Internal)?;
-
-        vm.memory_init(config.guest_memory.clone(), &kvm)
-            .map_err(StartMicrovmError::ConfigureVm)?;
-
-        // Instantiate the MMIO device manager.
-        // 'mmio_base' address has to be an address which is protected by the kernel
-        // and is architectural specific.
-        let mmio_device_manager = MMIODeviceManager::new(
-            config.guest_memory.clone(),
-            &mut (arch::MMIO_MEM_START as u64),
-            (arch::IRQ_BASE, arch::IRQ_MAX),
-        );
-
-        let mut vmm = Vmm {
-            stdin_handle: io::stdin(),
-            guest_memory: config.guest_memory,
-            vcpu_config: config.vcpu_config,
-            kernel_cmdline: config.kernel_cmdline,
-            vcpus_handles: Vec::new(),
-            exit_evt: None,
-            vm,
-            mmio_device_manager,
-            #[cfg(target_arch = "x86_64")]
-            pio_device_manager: PortIODeviceManager::new()
-                .map_err(Error::CreateLegacyDevice)
-                .map_err(StartMicrovmError::Internal)?,
-            write_metrics_event_fd,
-            seccomp_level: config.seccomp_level,
-            event_manager,
-        };
-
-        // For x86_64 we need to create the interrupt controller before calling `KVM_CREATE_VCPUS`
-        // while on aarch64 we need to do it the other way around.
-        #[cfg(target_arch = "x86_64")]
-        {
-            vmm.setup_interrupt_controller()?;
-            // This call has to be here after setting up the irqchip, because
-            // we set up some irqfd inside for some reason.
-            vmm.attach_legacy_devices()?;
-        }
-
-        // This was supposed to be the timestamp when the start command is recevied. Having this
-        // here just to create the vcpu; going forward the req timestamp will prob be somehow
-        // decoupled from the creation. At this point it's still fine because we create the
-        // builder and run the Vmm when the StartMicrovm request is received by the controller.
-        let request_ts = TimestampUs::default();
-        let vcpus = vmm.create_vcpus(config.entry_addr, request_ts)?;
-
-        #[cfg(target_arch = "aarch64")]
-        {
-            vmm.setup_interrupt_controller()?;
-            vmm.attach_legacy_devices()?;
-        }
-
-        Ok(VmmBuilderz { vmm, vcpus })
-    }
-
-    /// Return a reference to the guest memory object used by the builder.
-    pub fn guest_memory(&self) -> &GuestMemory {
-        self.vmm.guest_memory()
-    }
-
-    /// Returns a mutable reference to the guest kernel cmdline.
-    pub fn kernel_cmdline_mut(&mut self) -> &mut KernelCmdline {
-        &mut self.vmm.kernel_cmdline
-    }
-
-    /// Adds a MmioDevice.
-    pub fn attach_device(
-        &mut self,
-        id: String,
-        device: MmioDevice,
-    ) -> result::Result<(), StartMicrovmError> {
-        // TODO: we currently map into StartMicrovmError::RegisterBlockDevice for all
-        // devices at the end of device_manager.register_mmio_device.
-        let type_id = device.device().device_type();
-        let cmdline = &mut self.vmm.kernel_cmdline;
-
-        self.vmm
-            .mmio_device_manager
-            .register_mmio_device(self.vmm.vm.fd(), device, cmdline, type_id, id.as_str())
-            .map_err(StartMicrovmError::RegisterBlockDevice)?;
-
-        Ok(())
-    }
-
-    /// Start running and return the Vmm.
-    pub fn run(mut self, epoll_context: &mut EpollContext) -> result::Result<Vmm, VmmActionError> {
-        // Write the kernel command line to guest memory. This is x86_64 specific, since on
-        // aarch64 the command line will be specified through the FDT.
-        #[cfg(target_arch = "x86_64")]
-        kernel_loader::load_cmdline(
-            self.vmm.guest_memory(),
-            GuestAddress(arch::x86_64::layout::CMDLINE_START),
-            &self
-                .vmm
-                .kernel_cmdline
-                .as_cstring()
-                .map_err(StartMicrovmError::LoadCommandline)?,
-        )
-        .map_err(StartMicrovmError::LoadCommandline)?;
-
-        self.vmm.configure_system(self.vcpus.as_slice())?;
-
-        self.vmm.register_events(epoll_context)?;
-
-        self.vmm.start_vcpus(self.vcpus)?;
-
-        // Arm the log write timer.
-        // TODO: the timer does not stop on InstanceStop.
-        let timer_state = TimerState::Periodic {
-            current: Duration::from_secs(WRITE_METRICS_PERIOD_SECONDS),
-            interval: Duration::from_secs(WRITE_METRICS_PERIOD_SECONDS),
-        };
-        self.vmm
-            .write_metrics_event_fd
-            .set_state(timer_state, SetTimeFlags::Default);
-
-        // Log the metrics straight away to check the process startup time.
-        if LOGGER.log_metrics().is_err() {
-            METRICS.logger.missed_metrics_count.inc();
-        }
-
-        Ok(self.vmm)
-    }
 }
 
 /// Contains the state and associated methods required for the Firecracker VMM.
